Save another word on shorter types

+++ src/mr_protos.h
@@ -159,9 +159,9 @@ TYPEDEF_STRUCT (mr_array_param_t, ATTRIBUTES ( , "array parameters"),
 		)
 
 TYPEDEF_STRUCT (mr_bitfield_param_t, ATTRIBUTES ( , "bit-field parameters"),
-		(unsigned int, width, , "bit-field width in bits"),
-		(unsigned int, shift, , "bit-field shift in first byte"),
-		(ssize_t, size, , "size of bitfield array"),
+		(uint8_t, width, , "bit-field width in bits"),
+		(uint8_t, shift, , "bit-field shift in first byte"),
+		(unsigned, size, , "size of bitfield array"),
 		(uint8_t * , bitfield, , "flagged bit-fields saved as resizable array of bytes",
 		{ .offset = offsetof (mr_bitfield_param_t, size) }, "offset"), 
 		)
